requirements.txt
click
cssselect
Flask
itsdangerous
Jinja2
MarkupSafe
parsel
python-dateutil
PyYAML
selectorlib
six
w3lib
Werkzeug
